+++ integrationTests/e2e/setup.js
@@ -1,7 +1,9 @@
 const got = require('got');
 
 const vaultUrl = `${process.env.VAULT_HOST}:${process.env.VAULT_PORT}`;
-const vaultToken = `${process.env.VAULT_TOKEN || 'testtoken'}`
+const vaultToken = `${process.env.VAULT_TOKEN}` === undefined ? `${process.env.VAULT_TOKEN}` : "testtoken";
+
+const jsonStringMultiline = '{"x": 1, "y": "q\\nux"}';
 
 (async () => {
     try {
@@ -34,6 +36,44 @@ const vaultToken = `${process.env.VAULT_TOKEN || 'testtoken'}`
                     otherSecret: 'OTHERSUPERSECRET',
                 },
             }
+        });
+
+        await got(`http://${vaultUrl}/v1/secret/data/test-json-string`, {
+            method: 'POST',
+            headers: {
+                'X-Vault-Token': vaultToken,
+            },
+            json: {
+                data: {
+                    // this is stored in Vault as a string
+                    jsonString: '{"x":1,"y":"qux"}',
+                },
+            },
+        });
+
+        await got(`http://${vaultUrl}/v1/secret/data/test-json-data`, {
+            method: 'POST',
+            headers: {
+                'X-Vault-Token': vaultToken,
+            },
+            json: {
+                data: {
+                    // this is stored in Vault as a map
+                    jsonData: {"x":1,"y":"qux"},
+                },
+            },
+        });
+
+        await got(`http://${vaultUrl}/v1/secret/data/test-json-string-multiline`, {
+            method: 'POST',
+            headers: {
+                'X-Vault-Token': vaultToken,
+            },
+            json: {
+                data: {
+                    jsonStringMultiline,
+                },
+            },
         });
 
         await got(`http://${vaultUrl}/v1/sys/mounts/my-secret`, {
@@ -74,6 +114,18 @@ const vaultToken = `${process.env.VAULT_TOKEN || 'testtoken'}`
             json: {
                 foo: 'bar',
                 zip: 'zap',
+            },
+        });
+
+        await got(`http://${vaultUrl}/v1/secret/data/subsequent-test`, {
+            method: 'POST',
+            headers: {
+                'X-Vault-Token': vaultToken,
+            },
+            json: {
+                data: {
+                    secret: 'SUBSEQUENT_TEST_SECRET',
+                },
             },
         });
     } catch (error) {